move bolt redation logic to the logger service
