+++ internal/logger/logger.go
@@ -135,11 +135,13 @@ var sensitiveKeys = map[string]bool{
 	"auth_token": true,
 
 	// Connection details
-	"uri":      true,
-	"address":  true,
-	"host":     true,
-	"port":     true,
-	"bolt_uri": true,
+	"uri":        true,
+	"url":        true,
+	"listen_url": true,
+	"address":    true,
+	"host":       true,
+	"port":       true,
+	"bolt_uri":   true,
 }
 
 // IsSensitiveKey checks if a key contains sensitive information that should be redacted.
