+++ internal/logger/target.go
@@ -1,7 +1,7 @@
 // Copyright (c) "Neo4j"
 // Neo4j Sweden AB [http://neo4j.com]
 
-package database
+package logger
 
 import "net/url"
